Introduce ref and baseline batches (#3)
The final config.yaml must be generated first for a given number of
batches. For each batch a reference and baseline is done. The
optimisation draws a batch randomly for each trial.
The reasoning behind this is that the steplog output is large in
general but we might want to optimisae on more than only O(events)=10.

Co-authored-by: Benedikt Volkel <benedikt.volkel@cern.ch>